+++ src/main/java/net/geforcemods/securitycraft/screen/MineRemoteAccessToolScreen.java
@@ -25,7 +25,6 @@ import net.minecraft.world.level.block.Block;
 public class MineRemoteAccessToolScreen extends Screen {
 	private static final ResourceLocation TEXTURE = new ResourceLocation("securitycraft:textures/gui/container/mrat.png");
 	private static final ResourceLocation INFO_BOOK_ICONS = new ResourceLocation("securitycraft:textures/gui/info_book_icons.png"); //for the explosion icon
-	private final Component mratName = Utils.localize(SCContent.REMOTE_ACCESS_MINE.get().getDescriptionId());
 	private ItemStack mrat;
 	private Button[][] guiButtons = new Button[6][4]; //6 mines, 4 actions (defuse, prime, detonate, unbind)
 	private static final int DEFUSE = 0, ACTIVATE = 1, DETONATE = 2, UNBIND = 3;
@@ -35,7 +34,7 @@ public class MineRemoteAccessToolScreen extends Screen {
 	private final int[] lengths = new int[6];
 
 	public MineRemoteAccessToolScreen(ItemStack item) {
-		super(Component.translatable(item.getDescriptionId()));
+		super(item.getHoverName());
 
 		mrat = item;
 	}
@@ -45,14 +44,14 @@ public class MineRemoteAccessToolScreen extends Screen {
 		super.init();
 
 		int padding = 25;
-		int y = padding;
+		int y = 50;
 		int[] coords = null;
 		int id = 0;
 		int startX = (width - xSize) / 2;
 		int startY = (height - ySize) / 2;
 
 		for (int i = 0; i < 6; i++) {
-			y += 30;
+			y += 25;
 			coords = getMineCoordinates(i);
 
 			// initialize buttons
@@ -139,10 +138,10 @@ public class MineRemoteAccessToolScreen extends Screen {
 		RenderSystem._setShaderTexture(0, TEXTURE);
 		blit(pose, startX, startY, 0, 0, xSize, ySize);
 		super.render(pose, mouseX, mouseY, partialTicks);
-		font.draw(pose, mratName, startX + xSize / 2 - font.width(mratName), startY + -25 + 13, 0xFF0000);
+		font.draw(pose, title, startX + xSize / 2 - font.width(title) / 2, startY + 6, 4210752);
 
 		for (int i = 0; i < 6; i++) {
-			font.draw(pose, lines[i], startX + xSize / 2 - lengths[i] + 25, startY + i * 30 + 13, 4210752);
+			font.draw(pose, lines[i], startX + xSize / 2 - lengths[i] + 25, startY + i * 25 + 33, 4210752);
 		}
 	}