+++ Changelog.md
@@ -34,6 +34,7 @@
 - Fix: A player shooting a guardian with a taser gets damaged by the guardian's thorns
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser
 - Fix: Crashes when rendering some modded block entities in a Projector, or using them as a disguise
+- Fix: The Sentry-/Mine Remote Access Tool screens do not show the item's custom name
 
 [*] only available if the update_1_20 datapack is enabled